make service clients in main
